fix: RND-101: Fix ML backend compatibility with sdk>=1 (#570)
* fix: RND-101: Fix ML backend compatibility with sdk>=1

* Fix sdk version

* Install sdk from git

* Fix label-studio-tools imports

* Install numpy in requirements

* Cast float

* Add numpy reqs to easyocr and nemo

* Upgrade label-studio-sdk==1.0.2

---------

+++ label_studio_ml/examples/huggingface_ner/model.py
@@ -83,7 +83,7 @@ class HuggingFaceNER(LabelStudioMLBase):
                 entities = list(group)
                 start = entities[0]['start']
                 end = entities[-1]['end']
-                score = sum([entity['score'] for entity in entities]) / len(entities)
+                score = float(sum([entity['score'] for entity in entities]) / len(entities))
                 results.append({
                     'from_name': from_name,
                     'to_name': to_name,
